Add sgi setting process when ring is not empty
Add sgi setting process in do_icc_send and do_icc_perf
function when the ring checked is not empty.

The sig will be setted in icc_set_block in existing
processing, but in this case, it is possible that the
dest core cannot receive the last interrupt because
the last interrupt is being processed in dest core.

Signed-off-by: Chunguang Li <chunguang.li_1@nxp.com>